specs/modules/fatigue-analysis/stress_rainflow_to_damage/requirements.txt
# Core dependencies for Fatigue Analysis Module
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
pyyaml>=5.4.0
tqdm>=4.62.0

# Optional: Development dependencies
# pytest>=6.2.0
# pytest-cov>=2.12.0
# black>=21.6b0
# flake8>=3.9.0
# mypy>=0.910

# Optional: Monitoring dependencies
# schedule>=1.1.0
# watchdog>=2.1.0